requirements.txt
requests==2.29.0
beautifulsoup4
streamlit
numpy
Pillow
flask
keras~=2.15.0
tensorflow~=2.16